Fixed the Products problem, added sidebar
The fix was super easy, as I thought it was a query issue, didn't
realize how it was affecting me until my teacher pointed out the way the
LIMIT was added. She showed me a handy function for figuring out
problems like this (see line 47-50). Apparently the "DESC" and "LIMIT"
were right up against eachother in the output, which was easily fixed by
adding a space before the limit. AGH. PHP.