Fixed the Edit profile parser!
Changed the parser and just put it all into the edit page. Took out the
password because I don't believe there is a way (that I can do in a
timely matter and I'm short on time!) to make it work within one form.
It suggested to me by a friend that I should make the password changer
it's own form and not have the current password in the value. The next
step is adding the picture editor to the current edit profile form as
well as change password.